Add response body and headers as returnable data from GiphyService
