+++ packages/apihandler/test/giphy-service/giphy-service.test.ts
@@ -1,4 +1,5 @@
 import * as nock from 'nock';
+import { Headers } from 'node-fetch';
 import * as path from 'path';
 import { GiphyService } from '../../src/giphy-service/giphy-service';
 
@@ -29,22 +30,26 @@ describe('GiphyService', () => {
   describe('success', () => {
     const giphyService = new GiphyService(process.env[GIPHY_API_KEY_VARIABLE] || '');
 
-    describe('#getTrending', () => {
+    describe('#getSearch', () => {
       it('fetches trending', async () => {
-        const { nockDone } = await nockBack('trending.json', nockOptions);
-        const trending = await giphyService.getTrending();
-        expect(trending.data).toBeInstanceOf(Array);
-        expect(trending.data[0].url).toMatch(/https:\/\/giphy.com\/gifs\/.*/);
+        const { nockDone } = await nockBack('search.json', nockOptions);
+        const response = await giphyService.getSearch('ye');
+        const responseBody = await response.responseBody;
+        expect(responseBody.data).toBeInstanceOf(Array);
+        expect(responseBody.data[0].url).toMatch(/https:\/\/giphy.com\/gifs\/.*/);
+        expect(response.headers).toBeInstanceOf(Headers);
         nockDone();
       });
     });
 
-    describe('#getSearch', () => {
+    describe('#getTrending', () => {
       it('fetches trending', async () => {
-        const { nockDone } = await nockBack('search.json', nockOptions);
-        const trending = await giphyService.getSearch('ye');
-        expect(trending.data).toBeInstanceOf(Array);
-        expect(trending.data[0].url).toMatch(/https:\/\/giphy.com\/gifs\/.*/);
+        const { nockDone } = await nockBack('trending.json', nockOptions);
+        const response = await giphyService.getTrending();
+        const responseBody = await response.responseBody;
+        expect(responseBody.data).toBeInstanceOf(Array);
+        expect(responseBody.data[0].url).toMatch(/https:\/\/giphy.com\/gifs\/.*/);
+        expect(response.headers).toBeInstanceOf(Headers);
         nockDone();
       });
     });
@@ -52,9 +57,11 @@ describe('GiphyService', () => {
     describe('#getRandom', () => {
       it('fetches random with given tag', async () => {
         const { nockDone } = await nockBack('random.json', nockOptions);
-        const random = await giphyService.getRandom('kanye west');
-        expect(random).toEqual(expect.objectContaining({ data: expect.any(Object) }));
-        expect(random.data.embed_url).toMatch(/https:\/\/giphy.com\/embed\/.*/);
+        const response = await giphyService.getRandom('kanye west');
+        const responseBody = await response.responseBody;
+        expect(responseBody).toEqual(expect.objectContaining({ data: expect.any(Object) }));
+        expect(responseBody.data.embed_url).toMatch(/https:\/\/giphy.com\/embed\/.*/);
+        expect(response.headers).toBeInstanceOf(Headers);
         nockDone();
       });
     });
@@ -62,17 +69,21 @@ describe('GiphyService', () => {
     describe('#getTranslate', () => {
       it('fetches translate with given search term', async () => {
         const { nockDone } = await nockBack('translate.json', nockOptions);
-        const translate = await giphyService.getTranslate('yeezus');
-        expect(translate).toEqual(expect.objectContaining({ data: expect.any(Object) }));
-        expect(translate.data.images.downsized.url).toMatch(/https:\/\/media\d?.giphy.com\/media\/.*/);
+        const response = await giphyService.getTranslate('yeezus');
+        const responseBody = await response.responseBody;
+        expect(responseBody).toEqual(expect.objectContaining({ data: expect.any(Object) }));
+        expect(responseBody.data.images.downsized.url).toMatch(/https:\/\/media\d?.giphy.com\/media\/.*/);
+        expect(response.headers).toBeInstanceOf(Headers);
         nockDone();
       });
 
       it('fetches translate with given weirdness', async () => {
         const { nockDone } = await nockBack('translate_with_weirdness.json', nockOptions);
-        const translate = await giphyService.getTranslate('yeezus', 8);
-        expect(translate).toEqual(expect.objectContaining({ data: expect.any(Object) }));
-        expect(translate.data.images.downsized.url).toMatch(/https:\/\/media\d?.giphy.com\/media\/.*/);
+        const response = await giphyService.getTranslate('yeezus', 8);
+        const responseBody = await response.responseBody;
+        expect(responseBody).toEqual(expect.objectContaining({ data: expect.any(Object) }));
+        expect(responseBody.data.images.downsized.url).toMatch(/https:\/\/media\d?.giphy.com\/media\/.*/);
+        expect(response.headers).toBeInstanceOf(Headers);
         nockDone();
       });
     });
